Ping / Pong

+++ HelloHome.Central.Hub/MessageChannel/Messages/IncomingMessage.cs
@@ -6,7 +6,7 @@ namespace HelloHome.Central.Hub.MessageChannel.Messages
 	public class IncomingMessage : Message
 	{
 	    public int FromRfAddress { get; set; }
-	    public int Rssi { get; set; }
+	    public Int16 Rssi { get; set; }
 	}
 }
